Add git tag v0.1.0 to post-gen hook
Force the user to have a git tag made of the initial commit, because the
code is a "walking skeleton", aka the project structure is solid enough
to run, even if it doesn't do much.

This also means the CHANGELOG entry, currently flagging today's date as
v0.1.0, is now accurately backed with a git tag.