Add .cruft.json file, manually created
Despite the issues I spotted, the possibilities of Cruft for template
update are really huge, but we won't reap any of this template-update
value until we start keeping track of what versions of templates we use.

Let's assume cruft's issues can be fixed eventually, the file format is
adequate, so let's use it in the repo. We're defining this manually,
based on some assumptions on our part, rather than reusing cruft's code.