Change package_name to not be a template question
Stop asking the package_name with a default, instead hardcode it like we
always meant to.

Done by setting a file with that data in new vars/ folder, and using
"include" to bring back that data as if we had a "memory set" function.

Since we moved to vars/ folder, and file/folder names can't have
slash (/), use pathjoin func on file/folders instead to work around
missing char (trick from copier docs).