Add google-style docstring support in sphinx docs
Context:

In sphinx docs, the Napoleon plugin would support google-style
docstrings out of the box.

But default Sphinx is only using rST documents, which is a horrible
document-writing format. So we use myst, a markdown parser that replaces
rST with Markdown.

To make extracted API reference code understand markdown, we had
redirected the sphinx (napoleon) parser with the myst one. But that one
doesn't quite support fancy types.

So we had introduced sphinx-autodoc2, a variant of sphinx-autodoc that
is type-hint-aware, yey. But it didn't support google-style docstrings!

Solution:

Introduce a autodoc2 parser override that calls the Napoleon plugin
before rendering the sphinx docs, so we get all of:
- markdown not rST in sphinx docs
- type hints in rendered docs
- markdown in docstrings not just overall sphinx docs
- google-style docstrings to show arguments details and returns/raises

Add sample google-docstrings in generated code, showcasing the feature.